fix compilation and mark CreateCheckpoint as not usable

+++ alt/checkpoint.go
@@ -8,6 +8,8 @@ type Checkpoint struct {
 	ColShape
 }
 
+// Create a checkpoint
+// ! This method is unfinished in alt:V use at your own risk !
 func CreateCheckpoint(checkpointType uint8, x float32, y float32, z float32, radius float32, height float32, color RGBA) *Checkpoint {
 	ptr := C.core_create_checkpoint(C.ushort(checkpointType), C.float(x), C.float(y), C.float(z), C.float(radius), C.float(height), C.ushort(color.R), C.ushort(color.G), C.ushort(color.B), C.ushort(color.A))
 	checkpoint := &Checkpoint{}
@@ -30,12 +32,12 @@ func (c Checkpoint) Radius() float32 {
 
 func (c Checkpoint) Color() RGBA {
 	cRGBA := C.checkpoint_get_color(c.Ptr)
-	return RGBA{R:cRGBA.r,G:cRGBA.g,B:cRGBA.b,A:cRGBA.a}
+	return RGBA{R:uint8(cRGBA.r),G:uint8(cRGBA.g),B:uint8(cRGBA.b),A:uint8(cRGBA.a)}
 }
 
 func (c Checkpoint) NextPosition() Position {
 	cPos := C.checkpoint_get_next_position(c.Ptr)
-	return Position{X:cPos.x,Y:cPos.y,Z:cPos.z}
+	return Position{X:float32(cPos.x),Y:float32(cPos.y),Z:float32(cPos.z)}
 }
 
 func (c Checkpoint) SetCheckpointType(checkpointType uint8) {
